Fix namespacing weirdness
The Pronto runner gem strategy appears to be to name the gem like
`pronto-bundler_audit` when then leads to a directory structure of

  lib/pronto/bundler_audit.rb
  lib/pronto/bundler_audit/version.rb

But the namespacing that would be derived from those is in conflict.
For example, attempting to define

```ruby

module Pronto
  module BundlerAudit
    # ..
  end
end
```

... is in direct conflict with the fact that the bundler_audit gem
defines this same name as a class. So ruby fails trying to define it
as a module here.

To fix: Slightly change the module name as follows:

```ruby

module Pronto
  module BundlerAuditVersion
    # ..
  end
end
```

I took this pattern from the pronto-rubocop gem.