Add first round of tests
I really need to be able to stub constants to fully test this...
I've done what I can for now without this.

Create a separate Pronto::BundlerAudit::PatchHandler class which runs
Bundle Audit on the given patch and returns an Array of Pronto::Message
objects if any advisories are found.
- This is better than all being shoved into the base
  Pronto::BundlerAudit class.